Move mock objects from bia-ingest/test to bia-test-data
