+++ bia-ingest/pyproject.toml
@@ -13,6 +13,7 @@ requests = "^2.31.0"
 pytest = "^7.0"
 bia-shared-datamodels = { path = "../bia-shared-datamodels", develop = true }
 bia-integrator-api = { path = "../clients/python", develop = true }
+bia-test-data = { path = "../bia-test-data", develop = true }
 typer = "^0.12.3"
 typing-extensions = "^4.12.2"
 pydantic-settings = "^2.3.4"
