+++ bia-test-data/bia_test_data/__init__.py
@@ -1,0 +1,2 @@
+from pathlib import Path
+bia_test_data_dir = Path(__file__).parent.parent / "data"
