+++ bia-test-data/bia_test_data/mock_objects/mock_image_acquisition_protocol.py
