+++ bia-test-data/bia_test_data/mock_objects/mock_image_correlation_method.py
