+++ bia-test-data/bia_test_data/mock_objects/mock_specimen_imaging_preparation_protocol.py
