+++ bia-test-data/pyproject.toml
@@ -1,13 +1,16 @@
 [tool.poetry]
 name = "bia-test-data"
 version = "0.1.0"
-description = ""
+description = "Test data - biostudies submission, images and other artefacts"
 authors = ["Kola <kola@ebi.ac.uk>"]
+license = "Apache-2.0"
 readme = "README.md"
+packages = [{include = "bia_test_data"}]
 
 [tool.poetry.dependencies]
-python = "^3.11"
-
+python = "^3.10,<3.12"
+bia-shared-datamodels = { path = "../bia-shared-datamodels", develop = true }
+pydantic-settings = "^2.3.4"
 
 [build-system]
 requires = ["poetry-core"]
